Mention correct production build command in `README`
I noticed that the `jsBrowserProductionWebpack` command does not seem
to produce the `build/dist/js/productionExecutable` folder that the
`README` mentions.  While investigating this issue, I noticed that the
`publish.sh` script actually uses the `jsBrowserDistribution` command
instead.  This command produces the expected folder.  This leads me to
believe that the `README` is simply referencing the wrong command, and
this commit corrects this discrepancy.